Prek happy Bandit happy Init clean-up
