+++ library/pyproject.toml
@@ -262,6 +262,9 @@ project-excludes = [
     # main `.venv`; guarded at runtime with a clear ImportError if missing.
     "src/physicalai/policies/rldx1/model.py",
     "src/physicalai/policies/rldx1/components/backbone/modeling_qwen3_vl.py",
+    # xr0's Qwen3-VL backbone wrapper is dynamically typed (transformers modules
+    # resolve to `Module | Tensor`); same rationale as pi_gemma / qwen3_vl above.
+    "src/physicalai/policies/xr0/vlm.py",
 ]
