+++ library/src/physicalai/policies/xr0/config.py
@@ -13,6 +13,7 @@ from dataclasses import dataclass, field
 from typing import Literal
 
 from physicalai.config import Config
+
 from physicalai.data import Feature  # noqa: TC001 - Needed at runtime for type hint resolution
